requirements.txt
fireworks_ai==0.15.4
pydantic==2.9.2
streamlit==1.38.0
